move repos to models, remove table  for the extensions model
