+++ vault-plugins/flant_iam/backend/path_serviceaccount.go
@@ -204,7 +204,7 @@ func (b *serviceAccountBackend) handleExistence() framework.ExistenceFunc {
 		}
 
 		tx := b.storage.Txn(false)
-		repo := NewServiceAccountRepository(tx)
+		repo := model.NewServiceAccountRepository(tx)
 
 		obj, err := repo.GetById(id)
 		if err != nil {
@@ -233,7 +233,7 @@ func (b *serviceAccountBackend) handleCreate(expectID bool) framework.OperationF
 
 		tx := b.storage.Txn(true)
 		defer tx.Abort()
-		repo := NewServiceAccountRepository(tx)
+		repo := model.NewServiceAccountRepository(tx)
 
 		if err := repo.Create(serviceAccount); err != nil {
 			msg := "cannot create service account"
@@ -268,12 +268,12 @@ func (b *serviceAccountBackend) handleUpdate() framework.OperationFunc {
 		tx := b.storage.Txn(true)
 		defer tx.Abort()
 
-		repo := NewServiceAccountRepository(tx)
+		repo := model.NewServiceAccountRepository(tx)
 		err := repo.Update(serviceAccount)
-		if err == ErrNotFound {
+		if err == model.ErrNotFound {
 			return responseNotFound(req, model.ServiceAccountType)
 		}
-		if err == ErrVersionMismatch {
+		if err == model.ErrVersionMismatch {
 			return responseVersionMismatch(req)
 		}
 		if err != nil {
@@ -293,10 +293,10 @@ func (b *serviceAccountBackend) handleDelete() framework.OperationFunc {
 
 		tx := b.storage.Txn(true)
 		defer tx.Abort()
-		repo := NewServiceAccountRepository(tx)
+		repo := model.NewServiceAccountRepository(tx)
 
 		err := repo.Delete(id)
-		if err == ErrNotFound {
+		if err == model.ErrNotFound {
 			return responseNotFound(req, "service account not found")
 		}
 		if err != nil {
@@ -315,10 +315,10 @@ func (b *serviceAccountBackend) handleRead() framework.OperationFunc {
 		id := data.Get("uuid").(string)
 
 		tx := b.storage.Txn(false)
-		repo := NewServiceAccountRepository(tx)
+		repo := model.NewServiceAccountRepository(tx)
 
 		serviceAccount, err := repo.GetById(id)
-		if err == ErrNotFound {
+		if err == model.ErrNotFound {
 			return responseNotFound(req, model.ServiceAccountType)
 		}
 		if err != nil {
@@ -334,7 +334,7 @@ func (b *serviceAccountBackend) handleList() framework.OperationFunc {
 		tenantID := data.Get(model.TenantForeignPK).(string)
 
 		tx := b.storage.Txn(false)
-		repo := NewServiceAccountRepository(tx)
+		repo := model.NewServiceAccountRepository(tx)
 
 		list, err := repo.List(tenantID)
 		if err != nil {
@@ -348,157 +348,4 @@ func (b *serviceAccountBackend) handleList() framework.OperationFunc {
 		}
 		return resp, nil
 	}
-}
-
-type ServiceAccountRepository struct {
-	db         *io.MemoryStoreTxn // called "db" not to provoke transaction semantics
-	tenantRepo *TenantRepository
-}
-
-func NewServiceAccountRepository(tx *io.MemoryStoreTxn) *ServiceAccountRepository {
-	return &ServiceAccountRepository{
-		db:         tx,
-		tenantRepo: NewTenantRepository(tx),
-	}
-}
-
-func (r *ServiceAccountRepository) Create(serviceAccount *model.ServiceAccount) error {
-	tenant, err := r.tenantRepo.GetById(serviceAccount.TenantUUID)
-	if err != nil {
-		return err
-	}
-
-	if serviceAccount.Version != "" {
-		return ErrVersionMismatch
-	}
-	serviceAccount.Version = model.NewResourceVersion()
-	serviceAccount.FullIdentifier = model.CalcServiceAccountFullIdentifier(serviceAccount, tenant)
-
-	err = r.db.Insert(model.ServiceAccountType, serviceAccount)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-func (r *ServiceAccountRepository) GetById(id string) (*model.ServiceAccount, error) {
-	raw, err := r.db.First(model.ServiceAccountType, model.PK, id)
-	if err != nil {
-		return nil, err
-	}
-	if raw == nil {
-		return nil, ErrNotFound
-	}
-	serviceAccount := raw.(*model.ServiceAccount)
-	return serviceAccount, nil
-}
-
-/*
-TODO
-	* Из-за того, что в очереди формата TokenGenerationNumber стоит ttl 30 дней – token_ttl не может быть больше 30 дней.
-		См. подробнее следующий пункт и описание формата очереди.
-
-TODO Логика создания/обновления сервис аккаунта:
-	* type object_with_resource_version
-	* type tenanted_object
-	* validate_tenant(запрос, объект из стора)
-	* validate_resource_version(запрос, entry)
-	* пробуем загрузить объект, если объект есть, то:
-	* валидируем resource_version
-	* валидируем тенанта
-	* валидируем builtin_type_name
-	* если объекта нет, то:
-	* валидируем, что нам не передан resource_version
-*/
-func (r *ServiceAccountRepository) Update(serviceAccount *model.ServiceAccount) error {
-	stored, err := r.GetById(serviceAccount.UUID)
-	if err != nil {
-		return err
-	}
-
-	// Validate
-	if stored.TenantUUID != serviceAccount.TenantUUID {
-		return ErrNotFound
-	}
-	if stored.Version != serviceAccount.Version {
-		return ErrVersionMismatch
-	}
-	serviceAccount.Version = model.NewResourceVersion()
-
-	// Update
-
-	tenant, err := r.tenantRepo.GetById(serviceAccount.TenantUUID)
-	if err != nil {
-		return err
-	}
-	serviceAccount.FullIdentifier = model.CalcServiceAccountFullIdentifier(serviceAccount, tenant)
-
-	err = r.db.Insert(model.ServiceAccountType, serviceAccount)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-/*
-TODO
-	* При удалении необходимо удалить все “вложенные” объекты (Token и ServiceAccountPassword).
-	* При удалении необходимо удалить из всех связей (из групп, из role_binding’ов, из approval’ов и пр.)
-*/
-func (r *ServiceAccountRepository) Delete(id string) error {
-	serviceAccount, err := r.GetById(id)
-	if err != nil {
-		return err
-	}
-
-	return r.db.Delete(model.ServiceAccountType, serviceAccount)
-}
-
-func (r *ServiceAccountRepository) List(tenantID string) ([]string, error) {
-	iter, err := r.db.Get(model.ServiceAccountType, model.TenantForeignPK, tenantID)
-	if err != nil {
-		return nil, err
-	}
-
-	ids := []string{}
-	for {
-		raw := iter.Next()
-		if raw == nil {
-			break
-		}
-		u := raw.(*model.ServiceAccount)
-		ids = append(ids, u.UUID)
-	}
-	return ids, nil
-}
-
-func (r *ServiceAccountRepository) DeleteByTenant(tenantUUID string) error {
-	_, err := r.db.DeleteAll(model.ServiceAccountType, model.TenantForeignPK, tenantUUID)
-	return err
-}
-
-func (r *ServiceAccountRepository) Iter(action func(account *model.ServiceAccount) (bool, error)) error {
-	iter, err := r.db.Get(model.ServiceAccountType, model.PK)
-	if err != nil {
-		return err
-	}
-
-	for {
-		raw := iter.Next()
-		if raw == nil {
-			break
-		}
-		t := raw.(*model.ServiceAccount)
-		next, err := action(t)
-		if err != nil {
-			return err
-		}
-
-		if !next {
-			break
-		}
-	}
-
-	return nil
 }
