+++ vault-plugins/flant_iam/model/errors.go
@@ -1,0 +1,9 @@
+package model
+
+import "fmt"
+
+var (
+	ErrNotFound        = fmt.Errorf("not found")
+	ErrVersionMismatch = fmt.Errorf("version mismatch")
+	ErrAlreadyExists   = fmt.Errorf("already exists")
+)
